+++ lib/pages/create/complete_habit_dialog.dart
@@ -200,7 +200,6 @@ class _HabitDialogState extends State<HabitDialog> {
                       IconButton(
                         padding: const EdgeInsets.all(10),
                         icon: SvgPicture.asset('assets/icons/close.svg'),
-                        hoverColor: Colors.transparent,
                         onPressed: widget.beamerDelegate.beamBack,
                       ),
                     ],